+++ checkpoint/orbax/checkpoint/proto_checkpoint_handler.py
@@ -45,7 +45,7 @@ class ProtoCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
 
   async def async_save(
       self, directory: epath.Path, item: message.Message
-  ) -> Optional[List[future.Future]]:
+  ) -> List[future.Future]:
     """Saves the given proto.
 
     Args:
@@ -69,9 +69,8 @@ class ProtoCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
 
     async def async_save(directory, item):
       commit_futures = await self.async_save(directory, item)
-      if commit_futures:
-        for f in commit_futures:
-          f.result()
+      for f in commit_futures:
+        f.result()
 
     asyncio.run(async_save(directory, item))
     utils.sync_global_devices("ProtoCheckpointHandler:save")